requirements.txt
aiohttp==3.7.4.post0
anyio==3.2.1
anytree==2.8.0
apispec==4.7.0
apispec-webframeworks==0.5.2
argon2-cffi==20.1.0
arrow==1.1.1
asgiref==3.4.1
astroid==2.5.6
async-generator==1.10
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==3.3.1
branca==0.4.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
colorama==0.4.4
commonmark==0.9.1
cycler==0.10.0
dateparser==1.0.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
elasticsearch==7.13.1
entrypoints==0.3
Flask==2.0.1
Flask-Cors==3.0.10
flask-swagger-ui==3.36.0

geojson==2.5.0

gunicorn==20.1.0
idna==2.10
importlib-metadata==4.5.0

ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.9.1
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0

kiwisolver==1.3.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
marshmallow==3.12.1

mccabe==0.6.1
mistune==0.8.4
moment==0.12.1
multidict==5.1.0
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1

numpy==1.21.1
packaging==21.0
pandas==1.3.0
pandocfilters==1.4.3
parso==0.8.2

pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1

prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pylint==2.8.3
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.18.0
python-git-info==0.7.1
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==22.1.0
redis==3.5.3
regex==2021.8.3
requests==2.25.1
requests-unixsocket==0.2.0
rich==10.6.0
rope==0.19.0
scikit-learn==0.24.2
scipy==1.7.0

Send2Trash==1.7.1
six==1.16.0
sklearn==0.0
sniffio==1.2.0
tenacity==8.0.1
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
times==0.7
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.10.0.0
tzlocal==3.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
networkx==2.6.2
